+++ CHANGELOG.md
@@ -5,6 +5,12 @@
 - Project module
 - Task module
 
+## [0.0.2] - 2020-04-09
+### Added
+- Agile app create model `Value`
+- Agile app `value` admin
+- Agile app `value` fixtures
+
 ## [0.0.1] - 2020-04-09
 ### Added
 - Apache HTTP Server